+++ src/index.ts
@@ -1,4 +1,4 @@
 import { CHAIN_IDS } from "./constants/chains";
-import subscribeToLeverageTokenCreated from "./subscribers/leverageTokenCreatedSubscriber";
+import subscribeToLeverageTokenCreated from "./subscribers/leverageTokenCreated";
 
 subscribeToLeverageTokenCreated(CHAIN_IDS.BASE);
